+++ bot.py
@@ -17,7 +17,7 @@ if os.path.exists(".env"):
     load_dotenv()
 
 BOT_VERSION = "4.7.0"
-BOT_UPDATED = "2026-06-30"
+BOT_UPDATED = "2026-07-03"
 CHANGELOG_CHANNEL_ID = 1486967511839801414
 
 TOKEN = os.getenv("TOKEN")